pyproject.toml
[project]
name = "self-hosted-search-agent"
version = "0.1.0"
description = "Grounded search assistant: SearXNG + crawl4ai + EvidenceLoop"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "requests>=2.31.0",
    "chardet<6",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "crawl4ai>=0.6.0",
    "pydantic-ai>=1.70.0",
    "logfire>=4.30.0",
    "ddgs>=9.11.4",
]

[project.scripts]
search-agent = "search_agent.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["search_agent"]
